app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
scikit-learn
nltk
spacy
PyMuPDF
python-docx
pdfplumber
numpy
pandas
torch
transformers
joblib
gunicorn
python-dotenv
docx2txt
pdfminer.six